ref:新規作成と編集をリファクタ
