+++ src/components/hooks/useModal.tsx
@@ -8,19 +8,9 @@ export const useModal = () => {
   const closeModal = () => {
     setIsOpen(false)
   }
-  const [isOpenSecond, setIsOpenSecond] = useState(false)
-  const openModalSecond = () => {
-    setIsOpenSecond(true)
-  }
-  const closeModalSecond = () => {
-    setIsOpenSecond(false)
-  }
   return {
-    isOpen: isOpen,
-    openModal: openModal,
-    closeModal: closeModal,
-    isOpenSecond: isOpenSecond,
-    openModalSecond: openModalSecond,
-    closeModalSecond: closeModalSecond,
+    isOpen,
+    openModal,
+    closeModal,
   }
 }
