+++ src/components/model/resource/index.ts
@@ -1,0 +1,2 @@
+export * from "./Resource"
+export * from "./ResourceForm"
