+++ src/components/ui/Resource/ResourceForm.module.scss
@@ -1,15 +1,3 @@
-.header {
-  display: flex;
-  align-items: center;
-  justify-content: flex-start;
-}
-.close {
-  margin-left: auto;
-  cursor: pointer;
-}
-.close:hover {
-  color: #f9e547;
-}
 
 .button {
   background-color: #f9e547;
